+++ CHANGELOG.md
@@ -25,9 +25,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 -->
 
 ## [Unreleased](https://github.com/cyverse/troposphere/compare/v36-4...HEAD) - YYYY-MM-DD
+### Added
+  - Add instance count field in launch wizard to support multi-instance-launch
+    ([#832](https://github.com/cyverse/troposphere/pull/832))
 
-
-## [Unreleased](https://github.com/cyverse/troposphere/compare/v36-2...v36-4) - 2019-09-10
+## [v36-4](https://github.com/cyverse/troposphere/compare/v36-2...v36-4) - 2019-09-10
 ### Fixed
   - Patched style issue with XSEDE button
     ([#818](https://github.com/cyverse/troposphere/pull/818))